Add account mixer (#150)
* Add account mixer

This adds function to start/stop dcrwallet's account mixer and check if the
account mixer is active. SetAccountMixerConfig function was also added to
set what accounts will be used a mixed & change account.

* Add mix details to transactions struct

* txauthor: send regular transaction change to unmixed account if enabled

* Add account mixer notifications and stop mixer when balance reaches lowest split point

- stop mixers when canceling sync.
- check running mixers when a new block comes in and stop mixer if
  wallet outputs have been mixed.
- Move account mixer functions to multi-wallet to make use of shutdown context.
- Return error when attempting to start mixer with no mixable output.

* Fix lint

* Find account previously involved in cspp transactions

* Find last used cspp accounts

* Change FindLastUsedCSPPAccounts return type to string

* Fix compile error

* Fix lint

+++ address.go
@@ -69,7 +69,7 @@ func (wallet *Wallet) AddressInfo(address string) (*AddressInfo, error) {
 		if err != nil {
 			return nil, err
 		}
-		addressInfo.AccountNumber = accountNumber
+		addressInfo.AccountNumber = uint32(accountNumber)
 	}
 
 	return addressInfo, nil
